+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # ibello-api changelog
 
+## 1.21.1
+
+- New `JsonTransformer.registerDeserializer` method to fine-tune the JSON deserialization process
+
 ## 1.21.0
+
+2022-05-29
 
 - It is possible to load test data from CSV files
 - New `FeatureHandler.loadFeatures()` method